edge-rl-oncloud/demo/requirements_dashboard.txt
streamlit>=1.28.0
plotly>=5.17.0
paho-mqtt>=1.6.1
pandas>=2.0.0
